bruh

+++ src/main/java/com/finderfeed/solarforge/misc_things/RunicEnergy.java
@@ -22,7 +22,7 @@ public class RunicEnergy {
         CompoundTag nbt = playerEntity.getPersistentData();
         float maxEnergy = nbt.getFloat(MAX_ENERGY_TAG);
         if (maxEnergy == 0){
-            nbt.putInt(MAX_ENERGY_TAG,10000);
+            nbt.putFloat(MAX_ENERGY_TAG,10000);
             maxEnergy = nbt.getFloat(MAX_ENERGY_TAG);
         }
